displaygroups endpoint renamed to topology/actions. File name lowercase.
